+++ src/kit.ts
@@ -746,10 +746,7 @@ export async function scanForVSKits(pr?: ProgressReporter): Promise<Kit[]> {
         const sub_prs: Promise<Kit | null>[] = [];
         MSVC_HOST_ARCHES.forEach(hostArch => {
             targetArches.forEach(targetArch => {
-                const kit: Promise<Kit | null> = tryCreateNewVCEnvironment(inst, hostArch, targetArch, pr);
-                if (kit) {
-                    sub_prs.push(kit);
-                }
+                sub_prs.push(tryCreateNewVCEnvironment(inst, hostArch, targetArch, pr));
             });
         });
 
@@ -798,7 +795,7 @@ async function scanDirForClangForMSVCKits(dir: string, vsInstalls: VSInstallatio
         }
 
         const clangKits: Kit[] = [];
-        vsInstalls.forEach(async vs => {
+        for (const vs of vsInstalls) {
             const install_name = vsDisplayName(vs);
             const vs_arch = (version.target && version.target.triple.includes('i686-pc')) ? 'x86' : 'amd64';
 
@@ -829,7 +826,7 @@ async function scanDirForClangForMSVCKits(dir: string, vsInstalls: VSInstallatio
                     });
                 }
             }
-        });
+        }
         return clangKits;
     });
     return ([] as Kit[]).concat(...kits);